Update MainForm.cs
Changes in logic to find a compatible client. Previously there was only one OTB version to each Tibia version, with the new 7.x support there is more than one Tibia version to the same OTB version (710->730; 740->750 and 790 -> 792).
With theese changes, now the logics consider a list instead of a single client, it reads the list and try to match it by the dat and spr signature